Create Markov Chain Monte Carlo Simulator (Bayesian APC Model)
This is the code for an MCMC model used in projecting cancer incidence and mortality by making use of a Bayesian age-period-cohort model. This is inspired by a paper written by Isabelle Bray on “Application of Markov Chain Monte Carlo methods to projecting cancer incidence and mortality” making use of MCMC methods in epidemiological analytics.